image_analysis/requirements.txt
absl-py==1.4.0
anyio==3.7.0
astunparse==1.6.3
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.97.0
Flask==2.2.5
Flask-Cors==3.0.10
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.54.2
gunicorn==20.1.0
h11==0.14.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
jax==0.3.25
Jinja2==3.1.2
keras==2.11.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
python-multipart==0.0.6
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.7.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.22.0
uWSGI==2.0.21
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
